move mock namelist to translate
