+++ tests/translate/translate_pe_halo.py
@@ -1,8 +1,6 @@
-import fv3core.stencils.pe_halo as pe_halo
-from fv3core.decorators import StencilWrapper
 from fv3core.testing import TranslateFortranData2Py
-from fv3core.utils.grid import axis_offsets
 from fv3core.stencils.dyn_core import AcousticDynamics
+from fv3core.testing.translate import MockNamelist
 
 
 class TranslatePE_Halo(TranslateFortranData2Py):
@@ -22,14 +20,6 @@ class TranslatePE_Halo(TranslateFortranData2Py):
         }
         self.in_vars["parameters"] = ["ptop"]
         self.out_vars = {"pe": self.in_vars["data_vars"]["pe"]}
-
-        class MockNamelist:
-            def __init__(self) -> None:
-                self.hydrostatic = True
-                self.d_ext = 0
-                self.beta = 0
-                self.use_logp = False
-                self.convert_ke = True
 
         namelist = MockNamelist()
         acoustic_dynamics = AcousticDynamics(None, namelist, None, None, None)
